requirements.txt
langchain
python-dotenv
langchain-community
streamlit